create security for every class and owner

+++ src/main/java/com/creatorx/aslet/service/ClassService.java
@@ -1,17 +1,16 @@
 package com.creatorx.aslet.service;
 
 import com.creatorx.aslet.converter.ClassConverter;
-import com.creatorx.aslet.converter.ClassGroupConverter;
 import com.creatorx.aslet.dto.ClassCreateDto;
 import com.creatorx.aslet.dto.ClassDto;
-import com.creatorx.aslet.dto.ClassGroupCreateDto;
-import com.creatorx.aslet.dto.RequestMetadata;
 import com.creatorx.aslet.exception.ClassExistsException;
 import com.creatorx.aslet.exception.ClassNotFoundException;
+import com.creatorx.aslet.exception.NotBelongsToUserException;
 import com.creatorx.aslet.model.Class;
 import com.creatorx.aslet.model.ClassGroup;
 import com.creatorx.aslet.repository.ClassGroupRepository;
 import com.creatorx.aslet.repository.ClassRepository;
+import com.creatorx.aslet.utils.AccessUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -32,12 +31,12 @@ public class ClassService {
     private UserService userService;
 
     @Autowired
-    private RequestMetadata requestMetadata;
+    public AccessUtils accessUtils;
 
     public ClassDto createClass(ClassCreateDto classCreateDto) {
         if (classRepository.findByGradeAndLetter(classCreateDto.getGrade(), classCreateDto.getLetter()).size() > 0) throw new ClassExistsException();
         Class newClass = classConverter.classCreateDtoToClass(classCreateDto);
-        newClass.setOwner(userService.getUserByIdDefault(requestMetadata.getId()));
+        newClass.setOwner(userService.getUserByIdDefault(accessUtils.getUserId()));
         classRepository.save(newClass);
         ClassGroup classGroup = new ClassGroup();
         classGroup.setName(newClass.getClassName());
@@ -48,12 +47,14 @@ public class ClassService {
     }
 
     public List<ClassDto> getAllClasses() {
-        return classConverter.classToDto(classRepository.findAll());
+        return classConverter.classToDto(classRepository.findAllByOwner_Id(accessUtils.getUserId()));
     }
 
     public ClassDto getClassById(Long id) {
-        return classConverter.classToDto(classRepository.findById(id)
-                .orElseThrow(() -> new ClassNotFoundException(id)));
+        Class studentsClass = classRepository.findById(id)
+                .orElseThrow(() -> new ClassNotFoundException(id));
+        if (!accessUtils.isAdmin() && !accessUtils.doesBelongToUser(studentsClass.getOwner().getId())) throw new NotBelongsToUserException();
+        return classConverter.classToDto(studentsClass);
     }
 
     public Class getClassByIdDefault(Long id) {
@@ -62,20 +63,21 @@ public class ClassService {
     }
 
     public ClassDto updateClass(ClassDto updatedClass, Long id) {
+        Class studentsClass = classRepository.findById(id)
+                .orElseThrow(() -> new ClassNotFoundException(id));
+        if (!accessUtils.isAdmin() && !accessUtils.doesBelongToUser(studentsClass.getOwner().getId())) throw new NotBelongsToUserException();
         if (classRepository.findByGradeAndLetter(updatedClass.getGrade(), updatedClass.getLetter()).size() > 0) throw new ClassExistsException();
-        return classConverter.classToDto(
-                classRepository.findById(id)
-                .map(studentsClass -> {
-                    studentsClass.setGrade(updatedClass.getGrade());
-                    studentsClass.setLetter(updatedClass.getLetter());
-                    return classRepository.save(studentsClass);
-                }).orElseThrow(() -> new ClassNotFoundException(id)));
+        studentsClass.setGrade(updatedClass.getGrade());
+        studentsClass.setLetter(updatedClass.getLetter());
+        classRepository.save(studentsClass);
+        return classConverter.classToDto(studentsClass);
     }
 
     public void deleteClass(Long id) {
         if (!classRepository.existsById(id)) {
             throw new ClassNotFoundException(id);
         }
+        if (!accessUtils.isAdmin() && !accessUtils.doesBelongToUser(getClassByIdDefault(id).getOwner().getId())) throw new NotBelongsToUserException();
         classRepository.deleteById(id);
     }
 }
